+++ packages/components/src/components/Theme/index.ts
@@ -1,2 +1,2 @@
-export { RadixTheme } from './Theme';
+export { Theme, RadixTheme } from './Theme';
 export type { ThemeProps, AccentColor, GrayColor, Appearance, PanelBackground, Scaling, Radius } from './Theme';